KEYCLOAK-18369 Create MapKeycloakTransaction interface
